Refactored components/MainButtons
