+++ src/components/Chess.js
@@ -19,7 +19,7 @@ import GameTable from '../features/table/GameTable.js';
 import Board from '../features/Board.js';
 import Game from './Game.js';
 import HeuristicsBar from '../features/HeuristicsBar.js';
-import MainButtons from './MainButtons.js';
+import MainButtons from '../features/MainButtons.js';
 import SupportButtons from './SupportButtons.js';
 import '../index.css';
 import store from '../store';
